requirements.txt
pydantic>=2.0
pytest>=7.0 
black 
isort
flake8
httpx>=0.27.0  # Async HTTP client
python-dotenv  # For API key management 
pytest-asyncio>=0.23.0
anyio>=4.0.0
sqlalchemy>=2.0
tenacity
PyYAML
